Reflux without Promise present should work
Added tests for using Promises with and without (by monkey patching
utils as if the environment didn't have Promises). This also won't
require the tests to run without the promise polyfill.

The Reflux#createAction function now will use the triggerAsync instead
of triggerPromise if Promises is not present in the environment.

Refs #302